+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [2.0.0] - 2023-12-31
+
+* Added a settings page so users can specify the ACF local JSON save path.
+* Refactored the plugin to use classes and Composer’s autoloader.
+* Bumped “Tested up to” field to 6.4.2.
+
 ## [1.0.1] - 2022-09-18
 
 * Bumped “Tested up to” field to 6.0.2.